Move block_writer example from main Cockroach repo
The Block Writer example program is designed to write random blocks of data to a
single table in CockroachDB.  Writes are spread randomly across the table's
primary key via a randomized ID, and are never read.

The intention of this program is to create a large amount of data in CockroachDB
quickly, in order to quickly generate split and rebalance events.