Block Writer uses Postgres protocol
This commit modifies the block writer example to connect to cockroach using the
postgres wire protocol.

As an additional extra, this commit also allows the block writer to connect to a
postgres database. The sql dialect for postgres is different from that supported
by cockroach db, so usage of the 'sql-dialect' flag is necessary in order to
adjust the logic being used.