Adjust indexing for photos.comments table.
The commentsByPhotoID index was not actually unique as multiple comments
with the same timestamp could be inserted by the load-generator. Easy to
reproduce by running `photos --users 100`. Completely removed this index
and changed the primary key of the comments table to include the
timestamp which allowed a slight simplification of the query on the
comments table.