photos: Avoid inserting comments with empty photo IDs
Comments for the empty photo ID will have very high contention,
eventually bringing the process to a standstill.

Explicitly check for this case, and add a CHECK constraint to the schema
to ensure that any empty UUIDs that make it through to the database will
be rejected.

Updates cockroachdb/cockroach#9247