photos: add analytics queries to photos
Add analytics queries, which are optionally run with the
'--analytics-true' flag. This is default false to provide minimal
disruption to existing benchmarks, since it will cause throughput
numbers to decrease.

The analytics queries are specifically designed to trigger full table
scans and exercise various parts of the DistSQL code paths.